Devving with cobra commands
