+++ go.mod
@@ -5,6 +5,7 @@ go 1.14
 require (
 	github.com/hashicorp/go-plugin v1.4.5
 	github.com/privateerproj/privateer-sdk v0.0.0 // no release has been made yet
+	github.com/spf13/cobra v1.6.1 // indirect
 	golang.org/x/net v0.5.0 // indirect
 )
